paths.rs uses patterns now as well
