+++ crates/fuzzy_nucleo/src/paths.rs
@@ -9,12 +9,12 @@ use std::{
 use util::{paths::PathStyle, rel_path::RelPath};
 
 use nucleo::Utf32Str;
-use nucleo::pattern::{Atom, AtomKind, CaseMatching, Normalization};
+use nucleo::pattern::Atom;
 
 use fuzzy::CharBag;
 
 use crate::matcher::{self, LENGTH_PENALTY};
-use crate::{Cancelled, Case, positions_from_sorted};
+use crate::{Cancelled, Case, Query, count_case_mismatches, positions_from_sorted};
 
 #[derive(Clone, Debug)]
 pub struct PathMatchCandidate<'a> {
@@ -96,38 +96,7 @@ impl Ord for PathMatch {
     }
 }
 
-// Path matching is always case-insensitive at the nucleo level. `Case::Smart`
-// is honored as a *scoring hint*: when the query contains uppercase, candidates
-// whose matched characters disagree in case are downranked by a factor per
-// mismatch rather than dropped. This keeps `"Editor: Backspace"` matching
-// `"editor: backspace"` while still preferring exact-case hits.
 const SMART_CASE_PENALTY_PER_MISMATCH: f64 = 0.9;
-
-pub(crate) fn make_atoms(query: &str) -> Vec<Atom> {
-    query
-        .split_whitespace()
-        .map(|word| {
-            Atom::new(
-                word,
-                CaseMatching::Ignore,
-                Normalization::Smart,
-                AtomKind::Fuzzy,
-                false,
-            )
-        })
-        .collect()
-}
-
-// Only populated when we will actually charge a smart-case penalty, so the hot
-// path can iterate a plain `&[Atom]` and ignore this slice entirely.
-fn make_source_words(query: &str, case: Case) -> Option<Vec<Vec<char>>> {
-    (case.is_smart() && query.chars().any(|c| c.is_uppercase())).then(|| {
-        query
-            .split_whitespace()
-            .map(|word| word.chars().collect())
-            .collect()
-    })
-}
 
 fn case_penalty(mismatches: u32) -> f64 {
     if mismatches == 0 {
@@ -175,9 +144,7 @@ fn get_filename_match_bonus(
 
 fn path_match_helper<'a>(
     matcher: &mut nucleo::Matcher,
-    atoms: &[Atom],
-    source_words: Option<&[Vec<char>]>,
-    query_bag: CharBag,
+    query: &Query,
     candidates: impl Iterator<Item = PathMatchCandidate<'a>>,
     results: &mut Vec<PathMatch>,
     worktree_id: usize,
@@ -197,7 +164,6 @@ fn path_match_helper<'a>(
     let path_prefix_len = candidate_buf.len();
     let mut buf = Vec::new();
     let mut matched_chars: Vec<u32> = Vec::new();
-    let mut atom_matched_chars = Vec::new();
     let mut candidate_chars: Vec<char> = Vec::new();
     for candidate in candidates {
         buf.clear();
@@ -206,7 +172,7 @@ fn path_match_helper<'a>(
             return Err(Cancelled);
         }
 
-        if !candidate.char_bag.is_superset(query_bag) {
+        if !candidate.char_bag.is_superset(query.char_bag) {
             continue;
         }
 
@@ -219,70 +185,47 @@ fn path_match_helper<'a>(
 
         let haystack = Utf32Str::new(&candidate_buf, &mut buf);
 
-        if source_words.is_some() {
-            candidate_chars.clear();
-            candidate_chars.extend(candidate_buf.chars());
-        }
+        let Some(score) = query.pattern.indices(haystack, matcher, &mut matched_chars) else {
+            continue;
+        };
 
-        let mut total_score: u32 = 0;
-        let mut case_mismatches: u32 = 0;
-        let mut all_matched = true;
+        let case_mismatches = count_case_mismatches(
+            query.query_chars.as_deref(),
+            &matched_chars,
+            &candidate_buf,
+            &mut candidate_chars,
+        );
 
-        for (atom_idx, atom) in atoms.iter().enumerate() {
-            atom_matched_chars.clear();
-            let Some(score) = atom.indices(haystack, matcher, &mut atom_matched_chars) else {
-                all_matched = false;
-                break;
-            };
-            total_score = total_score.saturating_add(score as u32);
-            if let Some(source_words) = source_words {
-                let query_chars = &source_words[atom_idx];
-                if query_chars.len() == atom_matched_chars.len() {
-                    for (&query_char, &pos) in query_chars.iter().zip(&atom_matched_chars) {
-                        if let Some(&candidate_char) = candidate_chars.get(pos as usize)
-                            && candidate_char != query_char
-                            && candidate_char.eq_ignore_ascii_case(&query_char)
-                        {
-                            case_mismatches += 1;
-                        }
-                    }
-                }
-            }
-            matched_chars.extend_from_slice(&atom_matched_chars);
-        }
+        matched_chars.sort_unstable();
+        matched_chars.dedup();
 
-        if all_matched && !atoms.is_empty() {
-            matched_chars.sort_unstable();
-            matched_chars.dedup();
+        let length_penalty = candidate_buf.len() as f64 * LENGTH_PENALTY;
+        let filename_bonus = get_filename_match_bonus(&candidate_buf, &query.pattern.atoms, matcher);
+        let positive = (score as f64 + filename_bonus) * case_penalty(case_mismatches);
+        let adjusted_score = positive - length_penalty;
+        let positions = positions_from_sorted(&candidate_buf, &matched_chars);
 
-            let length_penalty = candidate_buf.len() as f64 * LENGTH_PENALTY;
-            let filename_bonus = get_filename_match_bonus(&candidate_buf, atoms, matcher);
-            let positive = (total_score as f64 + filename_bonus) * case_penalty(case_mismatches);
-            let adjusted_score = positive - length_penalty;
-            let positions = positions_from_sorted(&candidate_buf, &matched_chars);
-
-            results.push(PathMatch {
-                score: adjusted_score,
-                positions,
-                worktree_id,
-                path: if root_is_file {
-                    Arc::clone(path_prefix)
-                } else {
-                    candidate.path.into()
-                },
-                path_prefix: if root_is_file {
-                    RelPath::empty().into()
-                } else {
-                    Arc::clone(path_prefix)
-                },
-                is_dir: candidate.is_dir,
-                distance_to_relative_ancestor: relative_to
-                    .as_ref()
-                    .map_or(usize::MAX, |relative_to| {
-                        distance_between_paths(candidate.path, relative_to.as_ref())
-                    }),
-            });
-        }
+        results.push(PathMatch {
+            score: adjusted_score,
+            positions,
+            worktree_id,
+            path: if root_is_file {
+                Arc::clone(path_prefix)
+            } else {
+                candidate.path.into()
+            },
+            path_prefix: if root_is_file {
+                RelPath::empty().into()
+            } else {
+                Arc::clone(path_prefix)
+            },
+            is_dir: candidate.is_dir,
+            distance_to_relative_ancestor: relative_to
+                .as_ref()
+                .map_or(usize::MAX, |relative_to| {
+                    distance_between_paths(candidate.path, relative_to.as_ref())
+                }),
+        });
     }
     Ok(())
 }
@@ -296,13 +239,13 @@ pub fn match_fixed_path_set(
     max_results: usize,
     path_style: PathStyle,
 ) -> Vec<PathMatch> {
+    let Some(query) = Query::build(query, case) else {
+        return Vec::new();
+    };
+
     let mut config = nucleo::Config::DEFAULT;
     config.set_match_paths();
     let mut matcher = matcher::get_matcher(config);
-
-    let atoms = make_atoms(query);
-    let source_words = make_source_words(query, case);
-    let query_bag = CharBag::from(query);
 
     let root_is_file = worktree_root_name.is_some() && candidates.iter().all(|c| c.path.is_empty());
 
@@ -312,9 +255,7 @@ pub fn match_fixed_path_set(
 
     path_match_helper(
         &mut matcher,
-        &atoms,
-        source_words.as_deref(),
-        query_bag,
+        &query,
         candidates.into_iter(),
         &mut results,
         worktree_id,
@@ -352,9 +293,9 @@ pub async fn match_path_sets<'a, Set: PathMatchCandidateSet<'a>>(
         query.to_owned()
     };
 
-    let atoms = make_atoms(&query);
-    let source_words = make_source_words(&query, case);
-    let query_bag = CharBag::from(query.as_str());
+    let Some(query) = Query::build(&query, case) else {
+        return Vec::new();
+    };
 
     let num_cpus = executor.num_cpus().min(path_count);
     let segment_size = path_count.div_ceil(num_cpus);
@@ -371,8 +312,7 @@ pub async fn match_path_sets<'a, Set: PathMatchCandidateSet<'a>>(
                 .zip(matchers.iter_mut())
                 .enumerate()
             {
-                let atoms = atoms.clone();
-                let source_words = source_words.clone();
+                let query = &query;
                 let relative_to = relative_to.clone();
                 scope.spawn(async move {
                     let segment_start = segment_idx * segment_size;
@@ -389,9 +329,7 @@ pub async fn match_path_sets<'a, Set: PathMatchCandidateSet<'a>>(
 
                             if path_match_helper(
                                 matcher,
-                                &atoms,
-                                source_words.as_deref(),
-                                query_bag,
+                                query,
                                 candidates,
                                 results,
                                 candidate_set.id(),
